Separate component on createroom file.
