+++ jjikeobang-client/src/components/header/Logo.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import logoCircle from '../assets/img/logo-circle.png';
+import logoCircle from '../../assets/img/logo-circle.png';
 
 function Logo() {
     return (
