+++ jjikeobang-client/src/components/voteInfo/CandidateItem.js
