+++ jjikeobang-client/src/components/voteInfo/RoomHeader.js
