+++ jjikeobang-client/src/components/voteInfo/VoteStatusBoard.js
